prov/shm: change rx comp flags from smr flags to fi flags
The current flags passed into the rx completion functions are smr flags (uint16_t).
This is inconsistent with the tx completion function flags and requires an extra
translation in some places. Changing to fi flags (uint64_t) cleans up this definition
and prepares the functions and rx entry for easy integration with the shared CQ
and SRX changes coming.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>