prov/efa: move and rename rxr_pkt_req_max_data_size()
Function rxr_pkt_req_max_data_size() calculates the maximum
number of bytes of data can be saved in a REQ packet for
a given TX entry.

Because it is an operation on a tx_entry, this patch moved
the function to rxr_op_entry.c and renamed it to

    rxr_tx_entry_max_req_data_capacity()

The change from "size" to "capacity" is preparation for
an upcoming change.

Signed-off-by: Wei Zhang <wzam@amazon.com>