core/lock: Add ability to disable a genlock completely
The genlock supports a no-op option, which disables locking to
support a single-threaded environment.  However, it still uses
a mutex in debug mode to validate that the lock is being used
correctly.

This option will not work with the tcp rdm provider, as a higher
level lock is used to avoid nested locking issues further down.
The nested locks are converted to no-ops, but may be 'acquired'
recursively (which is why they are disabled).

Change the OFI_LOCK_NONE option to indicate that no locking is
needed and the debug checks should be disabled.  Add a new
OFI_LOCK_NOOP to indicate that locking is needed, but should be
a no-op.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>