prov/tcp: Fix incorrect signaling of the CQ
In tcpx_report_success, if the cq is associated with a wait object,
the cq is signaled using ofi_cq_signal().  This results in a narrow
race condition where a thread that has called fi_cq_sread() may
return without reading a completion before its timeout has expired.
The issue is that ofi_cq_signal() is also used by the app to wake-up
an thread blocked in fi_cq_sread().  However, in this case, the signal
was generated internally.

To fix, we simply need to signal the wait object directly, rather
than through the ofi_cq_signal() call itself.  This wakes-up the
thread to check for completions, but does not set the cq's wakeup
field, which indicates that the thread should return even if no
completions are available.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>