prov/net: Use progress lock in place of cq/domain locks
We hold the progress lock when writing completions to the cq
and checking for valid MRs when processing receives.  Use
the progress lock to protect the MR map and when reading from
the cq.  This eliminates the need to acquire separate locks
in the data transfer path (cq and domain for the mr map).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>