prov/net: Align provider name with configure name
Rename xnet directory to net and update build scripts to use
net instead of xnet.

The directory name and provider names need to match.  When
building a provider, the directory name is used with the
configure arguments (e.g. --enable-xnet).  This causes
confusion if the name differs from what the provider is
called (net in this case).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>