prov/cxi: Temp fix to allow compatibility with old CXI constants
Allow use of pre-upstream FI_ADDR_CXI and FI_PROTO_CXI values
to allow compatibility with the re-based official release tree
numeric values for those constants. This is a temporary fix to
provide a window of time for applications to operate without
having to recompile. This commit can not be pushed upstream
and should be reverted after code has been up-streamed for
some period of time.

NETCASSINI-3973

Signed-off-by: Steve Welch <welch@hpe.com>