core/pollfds: Add ability to manually select hot/cold sockets
The pollfds implementation maintains 2 arrays of fd's that it
monitors.  One is the default fd set of all watched fd's.  The
other is an array of active fd's.  The active fd set is disabled
by default, but can be enabled through the FI_POLL_FAIRNESS
variable.  An fd is moved to the active list if it had any events
on it within the last 'fairness' number of checks.  An fd is
removed from the active list if no events were reported.

In theory, this sounds like a great idea.  In practice, it does
not work well with some applications.

The issue is that there are situations where an application is
only trying to send or receive data from a small number of sockets,
maybe as few as 1, to the exclusion of all other traffic.  When
this occurs, sockets that have receive data sitting on them will
move to the active list and continue to generate events.  But
because the application isn't trying to receive data from the
corresponding peers (i.e. no tagged buffers are posted using
any source or the peer's source address), the sockets will always
remain on the active list.

The performance of poll depends on the number of sockets passed
to it and the number of events being reported.  When there are
many sockets generating events repeatedly, the application
performance suffers.  This is seen running MPI with a simple
pingpong test.  As the number of ranks increase, the pingpong
latency increases dramatically, event though only 2 ranks are
ever sending data to each other.  The other ranks are waiting in
the barrier phase, have sent data to the ranks running the pingpong
test, but the received data is not being handled.  The pending
data will be ignored until the pingpong ranks move to the barrier
phase.

Define a mechanism that allows a provider to directly specify if
a socket should be moved/removed from the active list.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>