fabtests/pytest: fix a bug in processing return code
Currently, in client/sever test, the code check returncode of
server first. If server's return is ENODATA and non-strict
mode is used, the code will report the test as skipped, without
checking the client's return code.

This is wrong because client might experienced a real failure
and has a real error code.

This patch addressed the issue by introudcing function

check_returncode_list()

which will check a list of return code, and report failure
when any return code in the list is a failure.

Signed-off-by: Wei Zhang <wzam@amazon.com>