prov/efa: recover unknown peer AH with raw address from efadv
It is possible that the EFA device would return an unknown AH(-1) for a known
peer in receive op. This is a bug and could cause the received packet to be
discarded. As a mitigation, the EFA provider attempts to recover the peer
address if the node supports extensible CQ using Raw Address:QPN:QKey.

Signed-off-by: wenduwan <wenduwan@amazon.com>