prov/efa: enable runting read with local RDMA read copy
Currently, runting read is enabled only when gdrcopy is available.
When gdrcopy is not available, long read is used.

This patch removed that restriction of runting read, so it will
be used even when GDRCOPY is not available, under which local
RDMA read copy is used with runting read.

This is because runting read with local RDMA read copy is faster
than long read.

Signed-off-by: Wei Zhang <wzam@amazon.com>