fabtests: Allow to use FI_ADDR_STR with -F
Some providers are not able to use IP addresses in any way, but
FI_ADDR_STR should be generic enough for those. When using this format
we must set the `service` argument to NULL.

Some tests did not take into account the addr_format in the hints, they
now do.

Signed-off-by: Quentin Boyer <quentin.boyer@eviden.com>