prov/opx: SDMA multiple packets per fill index
Utilize the driver's automatic header generation (AHG) to send multiple
packets in a single SDMA comp index queue entry. This enables us to send
more data concurrently via SDMA. This requires some changes to the way
our packet headers are structured when sending via SDMA.

PSN and origin TX
-----------------
The KDETH defines the PSN as a 32-bit field, but OPX uses a 24-bit PSN.
The remaining 8 bits are used to store the origin TX. However, the driver
assumes the full 32 bits are the PSN only. The driver also assumes the
PSN field is in network byte-order, and will convert it to host byte-order,
increment it, and then convert it back to network byte-order when generating
packet headers. This requires us to store the origin TX elsewhere in the
packet header, and convert the PSN to network byte-order when sending the
initial packet header, and converting the PSN to host-byte order when
receiving these packets. This only needs to be done for packets that may
have been sent via SDMA. To accommodate this, two macros have been created
for reading the PSN and origin TX from packets.

Target buffer address/offest
----------------------------
With AHG, the driver increments an offset field for each subsequent packet
header in the 6th qw of the packet header. Our dput packet headers don't
use offsets; instead, they use an address, in the 5th qw of the packet
header. So, it was necessary to move the dput address field to qw 6 (and
move the existing field in qw 6 to qw 5).

To further complicate things, the driver assumes the offset field is 4 bytes
in host byte-order, but the dput address field is 8 bytes. In order to get the
driver to correctly increment the 4-byte offset portion of the 8-byte address,
it's necessary to first convert the 8-byte address to network byte order (to
get the offset portion in the expected half of that qw), and then convert the
4-byte offset field back to host byte-order to be incremented. Then on the
receive side, the reverse is done to read the correct address value.

Delivery completion, max packets per write and iovecs
-----------------------------------------------------
When delivery completion is enforced (currently determined by the payload size),
we only need 2 iovecs: 1 for the request header, and one for the payload. The
current limit of 32 packets per SDMA write in this case is arbitrary. While not
limited by the driver, we still need a PSN and replay for each packet, so the
value should not be too large.

When delivery completion is not being used, the payload needs to be copied into
replay bounce buffers. Since the buffers are allocated with each individual
replay, it then becomes necessary for each packet being written to have its own
iovec. The driver enforces a limit of 8 data iovecs per SDMA request header.
As a result, we're limited to a max of 8 packets when not enforcing delivery
completion.

Signed-off-by: Ben Lynam <ben.lynam23@gmail.com>