prov/net: Add env var to rename provider
Since the net provider is a fork of the tcp provider (and will
eventually return to being the tcp provider), allow setting the
provider name through an environment variable.  This allows
testing the net provider as the "tcp" provider for apps that
explicitly request the provider by name.

This change is intended for debugging/testing purposes as
a work-around for running MPI with the net provider.  We're
currently needing to manually change the name in the code
and rebuild to use the provider (because of issues in MPI).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>