prov/net: Add override to disable auto-progress
The auto-progress thread can significantly impact performance
for applications that don't actually need auto-progress (e.g.
Intel MPI).  Allow forcing off auto-progress, even if the fi_info
attributes indicate that it should be enabled.

This provides a work-around for MPI, which turns on auto-progress,
but doesn't need it.  The impact is that additional threads are
spawned for each local MPI rank, with data bouncing between threads.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>