prov/net: Optimize single-threaded counter support
If the application is using FI_THREAD_DOMAIN, then all
accesses to counters will be serialized.  We don't need to
worry about the counter being updated while the app is
reading it.  This includes calls to wait on the counter.

This allows us to avoid spawning a progress thread just
because the apps wants suport for fi_cntr_wait().  Instead
fi_cntr_wait just needs to drive progress on the domain.
Avoiding the spawning of the progress thread can greatly
improve MPI performance.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>