prov/net: Rename 'internal' parameter used to clear progress signal
We pass an internal parameter to xnet_handle_event() to determine
if the calling thread should clear the event used to signal the
pollfds itself.  The signal wakes up any waiting thread to let
it know that the pollfds has been updated.  Currently, the only
thread that cares about this is the internal auto-progress thread.
No other calling threads touching the pollfds will ever block.

However, a subsequent patch will change this and allow the user's
thread to block on the pollfds in specific situations.  As a result,
the 'internal' parameter will be misnamed.  Change the name to
reflect that desired behavior, which is whether we clear the pollfds
signal or not.

We want to clear the signal if the calling thread is trying to
block on the pollfds.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>