core/genlock: Return true for nolock held assertion
If the genlock is set to nolock, we can't tell if the proper
higher-level locks are being held or not.  However, by failing
the ofi_genlock_held() call, we hit assertions in lower-level
util code, where it's impossible to locate the higher-level
lock.

For example, the net provider protects the CQ using its progress
lock.  But there are assertions in the util CQ code that the CQ
lock is held.  Either the net provider cannot use the util code
at all, forcing it to duplicate the functions, we disable the
checks completely, or we only disable them when nolock is set.

We pick the latter.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>