prov/net: Optimize getting completion counter
In xnet_get_cntr() we check the cq_flags in a series of if-else if
statements to retrieve the correct completion counter.  We can
instead locate the right counter directly by storing it when
setting the cq_flags.  This optimizes the successful counter path
by eliminating the if-else if branching.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>