prov/net: Fix cleanup of event_list
In xnet_close_conn(), we destroy any associated endpoint.  To
avoid accessing that endpoint, we need to remove any outstanding
asynchronous events against it.  However, we match any events
in the list if the rdm endpoint matches, not the msg endpoint.
As a result, we can remove unrelated events that we may need
to continue progressing other connections.

Fix the check to find events belonging only to the msg endpoint
being removed.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>