prov/net: Handle DISCONNECTED state in process_connreq
In the case of simultaneous connections with a peer, it's possible
for our connection to be rejected and closed before we can reach
the state of trying to connect.  In that case, the local ep will
be disabled and the connection marked as DISCONNECTED.  Since
this can occur, remove the assert() in xnet_process_connreq().

Signed-off-by: Sean Hefty <sean.hefty@intel.com>