prov/net: Combine rx_wait_list and rx_poll_list
Merge these two lists into one.  This will simplify the code, so
subsequent patches can optimize handling endpoints that are
actively waiting for receive buffers to be posted.

The rx_poll_list currently tracks which endpoints are active or
hot.  However, until a buffer is posted, no receive progress can
be made on the endpoint.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>