prov/net: Remove buffered data check from xnet_need_rx
xnet_need_rx returns true if there's buffered socket data to
read or we've read a receive message, but don't have a matching
receive buffer from the application.  The function should only
return true if we have an unmatched receive to process.

When we progress receives off the socket, we only leave data
buffered but unhandled if we hit some sort of error or the
prior message was unmatched.  See the do-while loop in
xnet_progress_rx():

	do {
		if (ep->cur_rx.hdr_done < ep->cur_rx.hdr_len) {
			ret = xnet_recv_hdr(ep);
		} else {
			ret = ep->cur_rx.handler(ep);
		}

	} while (!ret && ofi_bsock_readable(&ep->bsock));

If ret != -FI_EAGAIN, this error is fatal to the endpoint and
we disable it.  If ret == FI_EAGAIN, we're dealing with an
unmatched receive (i.e. unexpected message).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>