prov/net: Remove xnet_need_rx() checks from CM path
The CM data exchange is done by send and receiving directly
on the socket, bypassing the buffered socket implementation.
As a result, we have no data in the buffered socket, nor have
we tried to process any non-CM messages.

This removes unneeded code and simplifies the use of xnet_need_rx
for subsequent changes.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>