prov/net: Move check for default Rx completion to message handling
When receive buffers are posted to an SRX, we don't know if it
will generate a CQ entry until a specific endpoint has claimed
the buffer.  As a result, we need to update the recv_entry->
cq_flags at that point.  To simplify the code, always perform
the check to see if an endpoint should generate a CQ entry by
default after claiming a receive buffer.  This removes the
need to perform the check when posting a receive buffer directly
to the endpoint (e.g. in xnet_recv()) and simplifies the logic.

This change will make it easier to merge the tagged and untagged
message handling by the progress engine.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>