prov/net: Set ep is_active when adding socket to poll set
Sockets are automatically placed on the hot list when added
to a poll set.  Set is_active to true when adding the socket
to ensure it is in sync with whether the socket is hot or not.

And to better clarify the association here, rename is_active
to is_hot.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>