core/pollfds: Update hot fds when modifying pollfds
ofi_pollfds_mod() updates the main poll fd array, but does not
modify the hot fd array.  As a result, any update will only be
seen when polling the full array.

This problem can lead to significant drops in performance when
running pingpong or bandwidth tests once transfers hit beyond
a given size.  The issue is that the active socket used for the
transfer gets EAGAIN back from the socket send() call because
of a lack of local or remote buffer space to receive the data.
The net provider udpates the socket to check for the POLLOUT
event, however, that update is only applied to the full poll
fd array, and not the hot fd array.  When FI_POLL_FAIRNESS is
set to a high value, it takes that many iterations through the
progress engine before the next send can occur.  This cycle
continues.

As part of the change, automatically move a socket to the hot
set if it's been modified, except in the case where all events
for the socket have been cleared.  In that case, remove it from
the hot set.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>