prov/efa: release tx_entry that written CQ error for RNR
For an tx_entry that encountered RNR and wrote an CQ
error entry for it, this patch release the tx_entry
when it does not have outstanding TX ops.

The tx entry cannot be released when it has outstanding TX
ops because the device will report completions for
those outstanding TX ops, and tx_entry is needed to
process the completions.

Signed-off-by: Wei Zhang <wzam@amazon.com>