prov/rxm: Hold AV lock throughout AV remove
The av remove invokes a series of functions which separately
acquire and release the AV lock.  As an optimization, hold the AV
lock throughout the entire process.

This change also allows rxm to handle the case where an address
is inserted into the AV more than once.  The util_av code
already supports this by referencing counting on the AV entry.
However, rxm always removed the associated peer_addr when remove
is called, even if the av entry remained valid.  The restructured
code leaves the peer_addr with the av_entry as long as the entry
is still valid.

(Note that support for duplicate insertions is not guaranteed by
the API.  However, this avoids a crash in the case that it does
occur and helps with providing debugging output for the app.)

Signed-off-by: Sean Hefty <sean.hefty@intel.com>