fabtests/rdm_atomic: simplify test to use more common code
The previous version of the test skipped the MR registration because
it had two more buffers to register. However, the registration of the
first buffer should be the same as the common code (since it checks the
mr_local and RMA flags and registers them accordingly).
This simplifies the test to just call ft_init_fabric which will register
the normal buffer. Then the test can register its additional two buffers
as well using the common ft_reg_mr call.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>