prov/net: Implement hot polling list directly
Maintain a separate poll set for hot sockets, rather than
using the pollfds hot fd feature.  This will be used later to
mix epoll for the full socket list with poll for the hot list.

The hot feature for the pollfds will be removed, as it does
not work without direct involvement of the user.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>