prov/net: Move sockets off of hot set quicker
We currently check if a socket has gone cold only once every
poll fairness iterations.  This can leave sockets as hot for
longer than necessary.  For example, in an MPI alltoall test
with 64 ranks, the average number of sockets on the hot list
was around 60 with fairness set to 100.  However, with fairness
set to 10, the average number of hot sockets dropped to 2.

Because the performance of poll is related to the number of
sockets being polled, a large number of sockets on the hot
list limits the usefulness of having the hot list.

Change how sockets are moved off of the hot list in order to
remove sockets that may have received a single message from
the list quicker.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>