core and prov/shm: Implement and use ipc_cache.
Opening and closing the HMEM IPC handles are expensive operations. They
should not be performed on every data operation.

This patch adds a caching mechanism. On the very first IPC handle open
call, the address is cached and the mapped memory address is stored. In
subsequent calls the cache is queried and the mapped address is retrieved.

The cache is created as an object of ofi_mr_cache, with the following component:
cache->add_region: open the ipc handle
cache->delete_region: close the ipc handle
cache->monitor: the created ipc monitor.

Signed-off-by: Shi Jin <sjina@amazon.com>