prov/opx: Add work pending completion queue to enable more parallel SDMA operations
Add a low-priority work pending queue to handle deferred work items
where the work has already been done and we're just waiting for
confirmation before cleanup. Having a separate queue for this allows
us to begin work on the next item in the main work pending queue while
we wait.

Signed-off-by: Ben Lynam <ben.lynam23@gmail.com>