fabtests/common, fi_scalable_ep: reorganize ft_alloc_msgs calls
Change the ordering of the allocation of messages during initiailization
so that the messages are allocated after EPs are created and enabled.
This is in preparation of the FI_MR_ENDPOINT support.

This requires modifying the fi_scalable_ep and ft_shared_ctx tests to
manually call ft_alloc_msgs between enabling all of the endpoints and
posting all the receives.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>