contrib/intel/jenkins: Add a functional summary stage to Jenkinsfile
This stage is only viewable on passed pipelines. This is because jenkins
skips future stages if earlier stages failed. So, this stage is just to
provide a clear place to look on passed pipelines for the functional summary.
The functional summary will still be provided in the post stage inside of the
always{} clause incase of a test failure.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>