prov/shm: refactor rx completion functions
All rx completion instances are called by the parent function smr_complete_rx.
This simplifies the code so that smr_complete_rx gets the smr_cq and fiaddr and
passes them in instead of the function pointer doing that. Since the shm id->fiaddr
lookup is contanst, it is trivial to always pass it in.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>