prov/shm: use fi_peer_rx_entry
Refactor the smr_rx_entry to internally use the fi_peer_rx_entry
and its fields. This prepares for the srx changes to come.

This also modifies the address lookup when matching receive entries
from using the internal shm id to using the fiaddr to match.
Shm uses an internal address table so looking up the fiaddr is trivial.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>