fabtests/multi_recv: Fix possible crash in cleanup
The multi_recv test frees the mr_multi_recv memory region and
associated data buffers while they are still posted to an active
endpoint.  This can lead to a crash in the provider trying to
access the buffer, and occurs with the net provider as a result
of enabling the auto-progress thread.

Replace the in-band synchronization message with an out of
band message that safely indicates there is no active
traffic on the endpoint.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>