prov/net: Report user buffer with multi-recv completions
When reporting a completion that landed in a multi-receive
buffer, we need to include the offset into the buffer where
the message landed.  Store this with the receive xfer_entry
and return it with the completion.

To make things simpler and for stronger debugging, we
always write the receive buffer for rx completions, even
for cases when it is not necessary.  Likewise, we record
the user buffer on transmits.

These changes will allow for subsequent patches to analyze
the buffer contents between posting and completions and
verification at the receiver for data verification.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>