prov/net: Modify FI_PEEK check to examine pending receives
MPI probe will call fi_trecvmsg() with the FI_PEEK flag set.
Because the net provider always fails the FI_PEEK call,
applications waiting on MPI iprobe to return that a message
is ready to be received before it posts the receive buffer can hang.

Update peek to check against any messages that may be queued
for receiving.  This results in the net provider behaving as
if it had an unexpected message queue of size 1.

This also fixes generating a completion in the fi_recvmsg() calls.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>