prov/verbs: Add option to disable dmabuf support
Currently the verbs provider uses dmabuf based memory registration
whenever possible and only falls back to regular memory registration
when the first approach fails. This works well most of the time.
However, sometimes it may be desirable to disable the dmabuf based
registration. For example, to test the functionality of the dmabuf
peer-mem adapter driver.

This patch adds a runtime option to the verbs provider that allows
explicit enabling/disabling of dmabuf support.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>