contrib/intel: change node name delimiter from - to _
The Jenkins nodes use the naming convention "node-fabric"
Jenkins parses by the '-' and assumes the first element to be the
node name to ssh into. This delimiter conflicts with normal node
naming convention which uses dashes within the node name.
This changes the node/fabric delimiter from - to _ so it doesn't
conflict with regular node names.

Note that this will require a momentary down period between naming
the nodes and merging the PR.
After this change is merged, PRs will have to rebase on upstream
in order to pass.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>