prov/shm: fix ZE IPC race condition on startup
ZE IPC needs a specific handle sharing protocol in shm and
needs a Unix domain socket to share fds. The IPC enable check
makes sure that the peer has opened this Unix fd and can
exchange the shared handles.
This fixes a race condition where the peer has not initialized its
socket and fails that check but the sender tries to send an IPC
message. In this case, the sender would go to the do_ipc case
but then fail the ZE check and try to send a generic IPC message
which would fail because ZE needs its special protocol.
A ZE IPC send should never take the generic IPC protocol - it should
take the ZE IPC protocol if possible and fallback to SAR if necessary.
This allows the fallback to happen

Cherry-picked from commit e4ac7d0dc0608fce65d3e9cb1b1963ddceecd97a

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>